feat: wire Verus verification for aspen-cluster and aspen-transport
Integrate cluster and transport Verus specs into automated verification pipeline.

- Add cluster and transport to nix run .#verify-verus (now 5/5 crates)
- Add verify-verus-cluster and verify-verus-transport apps
- Fix Verus syntax issues in cluster and transport specs
- Update CI verus-check to include all 5 crates
- All quick syntax checks now pass

+++ crates/aspen-cluster/verus/trust_spec.rs
@@ -313,9 +313,9 @@ verus! {
         current_time_ms: u64,
     ) -> u64 {
         if current_time_ms >= received_at_ms {
-            (current_time_ms - received_at_ms) / 1000
+            ((current_time_ms - received_at_ms) / 1000) as u64
         } else {
-            0  // Clock went backward, treat as fresh
+            0u64  // Clock went backward, treat as fresh
         }
     }
